src/requirements.txt
# requirements.txt
# Dependencies for the Chimaera ECE

# Core agent framework
Elysia

# For loading configuration from .env files (good practice for API keys)
python-dotenv

# For parsing the config.yaml file
pyyaml

# For interacting with the Neo4j knowledge graph
neo4j

# For local LLM serving
ollama
pydantic
